+++ dart/lib/src/noop_client.dart
@@ -5,6 +5,7 @@ import 'protocol.dart';
 import 'scope.dart';
 import 'sentry_options.dart';
 import 'transport/transport.dart';
+import 'stack_trace.dart';
 
 class NoOpSentryClient implements SentryClient {
   NoOpSentryClient._();
@@ -25,11 +26,21 @@ class NoOpSentryClient implements SentryClient {
   Transport transport;
 
   @override
-  Future<SentryId> captureEvent(SentryEvent event, {stackFrameFilter, scope}) =>
+  Future<SentryId> captureEvent(
+    SentryEvent event, {
+    StackFrameFilter stackFrameFilter,
+    Scope scope,
+    dynamic hint,
+  }) =>
       Future.value(SentryId.empty());
 
   @override
-  Future<SentryId> captureException(throwable, {stackTrace, scope}) =>
+  Future<SentryId> captureException(
+    dynamic throwable, {
+    dynamic stackTrace,
+    Scope scope,
+    dynamic hint,
+  }) =>
       Future.value(SentryId.empty());
 
   @override
@@ -39,6 +50,7 @@ class NoOpSentryClient implements SentryClient {
     String template,
     List<dynamic> params,
     Scope scope,
+    dynamic hint,
   }) =>
       Future.value(SentryId.empty());